refactor: Lotto#getPrizeCondition()의 인자 변경-조금 더 캡슐화
LottoTest 테스트코드 수정하면서 Lotto 클래스 리팩토링
