+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,8 +1,5 @@
-package lotto;
+package lotto.domain;
 
-import lotto.domain.BonusNumber;
-import lotto.domain.Lotto;
-import lotto.domain.WinningNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -26,17 +23,17 @@ class LottoTest {
                 .isInstanceOf(IllegalArgumentException.class);
     }
 
-    @DisplayName("로또 넘버 비교하기")
+    @DisplayName("로또 넘버 비교하기 (5개 번호가 같고, 보너스 번호가 다르면 3등이다.)")
     @Test
     void compareNumbers() {
         // given
         Lotto lotto = new Lotto(List.of(1, 2, 3, 4, 5, 6));
         WinningNumbers winningNumbers = WinningNumbers.create("1,2,3,4,5,7");
-        BonusNumber bonusNumber = BonusNumber.create("1");
+        BonusNumber bonusNumber = BonusNumber.create("8");
+        WinningLotto winningLotto = WinningLotto.create(winningNumbers, bonusNumber);
 
         // when, then
-        lotto.getPrizeCondition(winningNumbers, bonusNumber);
-//        assertThat(lotto.containsBonusNumber(1)).isTrue();
-//        assertThat(lotto.countMatchingNumbers(winningNumbers)).isEqualTo(5);
+        PrizeCondition prizeCondition = lotto.getPrizeCondition(winningLotto);
+        assertThat(prizeCondition).isEqualTo(PrizeCondition.THIRD);
     }
 }
